server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
git+https://github.com/Lu-Yi-Hsun/iqoptionapi.git
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3

